requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
dj-database-url==0.5.0
Django==3.2.6
django-embed-video==1.4.0
django-heroku==0.3.1
gunicorn==20.1.0
idna==3.2
psycopg2==2.9.1
python-decouple==3.4
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.3.0
wincertstore==0.2
